fix: split usdc in common.go and contracts.go
